refactor: move PG DB preparation when using e2e tests split
