+++ .github/workflows/ci.yml
@@ -35,11 +35,12 @@ jobs:
         with:
           fetch-depth: 0
 
-      - run: pwd
-
       # TODO: ensure vars are shared with agents --with-env-vars="auto"
-      - run: npx nx-cloud start-ci-run --distribute-on="3 linux-medium-js" --stop-agents-after="e2e-ci"
-        if: ${{ !env.ACT }}
+
+      # --require-explicit-completion because Nx agents might not receive tasks for more than 30 seconds when preparing the e2e test environment
+      # @see https://nx.dev/ci/troubleshooting/ci-execution-failed#the-nx-cloud-heartbeat-process-failed-to-report-its-status-in-time
+      - run: npx nx-cloud start-ci-run --distribute-on="3 linux-medium-js" --stop-agents-after="e2e-ci" --require-explicit-completion
+        if: ${{ env.NX_CLOUD_DISTRIBUTED_EXECUTION == 'true' }}
 
       - uses: actions/setup-node@v4
         with:
@@ -74,7 +75,19 @@ jobs:
       - run: npx nx-cloud record -- docker compose -p cat-fostering logs keto -n 500
       - run: npx nx-cloud record -- docker compose -p cat-fostering logs kratos -n 500
 
-      # - run: npx nx affected -t e2e --no-agents
+      - run: npx nx affected -t e2e --no-agents
+        if: ${{ env.NX_CLOUD_DISTRIBUTED_EXECUTION == 'false' }}
 
-      # to enable DTE for split e2e tests, use the following command instead and set stop-agents-after="e2e-ci"
-      - run: npx nx affected -t e2e-ci
+      # to enable DTE for e2e tests splitting, use the following command instead and set stop-agents-after="e2e-ci"
+      - run: |
+          npm run ory:generate:keto -- -e .env.ci
+          docker compose -p cat-fostering restart keto
+          npm run ory:generate:kratos -- -e .env.ci
+          docker compose -p cat-fostering restart kratos
+          npm run pg:create:connection -- -e .env.ci
+          docker compose -p cat-fostering restart api
+          npx nx affected -t e2e-ci
+        if: ${{ env.NX_CLOUD_DISTRIBUTED_EXECUTION == 'true' }}
+
+      - run: npx nx-cloud complete-ci-run
+        if: always() && ${{ env.NX_CLOUD_DISTRIBUTED_EXECUTION == 'true' }}
