+++ apps/cat-fostering-api-e2e/src/support/global-teardown.ts
@@ -15,16 +15,22 @@ const dockerEnvPath = join(__dirname, '..', '..', '..', '..', '.env');
 export default async (): Promise<void> => {
   console.log(globalThis.__TEARDOWN_MESSAGE__);
 
-  await (globalThis.__DB_CONNECTION__ as DataSource)?.destroy();
-
-  if (process.env.CI && !process.env.ACT) {
+  if (
+    process.env.CI ||
+    process.env.ACT ||
+    // unfortunately, the task generated for tests splitting do not contain the target name
+    process.env.NX_TASK_TARGET_TARGET === 'e2e-ci'
+  ) {
     return;
   }
-  generateOryKetoConfig(dockerEnvPath);
-  restartService('keto');
 
-  generateOryKratosConfig(dockerEnvPath);
-  restartService('kratos');
+  if (process.env.NX_TASK_TARGET_TARGET === 'e2e') {
+    await (globalThis.__DB_CONNECTION__ as DataSource)?.destroy();
+    generateOryKetoConfig(dockerEnvPath);
+    restartService('keto');
+    generateOryKratosConfig(dockerEnvPath);
+    restartService('kratos');
+  }
 };
 
 cleanupRegisteredPaths();
