+++ libs/pg-config/README.md
@@ -1,0 +1,3 @@
+# pg-config
+
+This library was generated with [Nx](https://nx.dev).
