+++ tsconfig.base.json
@@ -28,7 +28,8 @@
       "@cat-fostering/nestjs-utils": ["libs/shared/nestjs-utils/src/index.ts"],
       "@cat-fostering/ory-config-generators": [
         "libs/ory-config-generators/src/index.ts"
-      ]
+      ],
+      "@cat-fostering/pg-config": ["libs/pg-config/src/index.ts"]
     },
     "typeRoots": ["node_modules/@types", "types"],
     "types": ["@dotenvx/dotenvx"]
